+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- silo-core: fix leverage bug with swap abusing debt share token receive allowance and ERC-20 approval
 
 ### [3.8.0] - 2025-06-26
 ### Added